requirements.txt
fastapi>=0.115.2
uvicorn[standard]>=0.30.1
ultralytics>=8.3.0
python-multipart>=0.0.18
prometheus-client>=0.20.0
Pillow>=10.3.0
opencv-python-headless>=4.10.0
onnxruntime>=1.18.0
gradio>=6.0.0
audioop-lts>=0.2.0 ; python_version >= "3.13"
